fix bugs

+++ UDRoute/ManagedSHA256.cs
@@ -26,7 +26,7 @@ namespace UDRoute
                 0x510e527f, 0x9b05688c, 0x1f83d9ab, 0x5be0cd19
             };
 
-            int bitLen = message.Length * 8;
+            ulong bitLen = (ulong)message.Length * 8;
             int padLen = 64 - ((message.Length + 8) % 64);
             if (padLen < 1) padLen += 64;
